+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 ### Fixed
+- Highlight matched side bar button when collapsed
 - Use scrollable list view in firmwares listing
 
 ## [0.1.6] - 2021-05-05