+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add copyable label with transaction hash to the end of send flow
 
 ### Changed
+- Update GDK to 0.0.42
 
 ### Fixed
 - Highlight matched side bar button when collapsed